Move Statusbar to own view, made into prism region.
