+++ Framework/Wider.Core.Wpf/Views/StatusBar.xaml
@@ -1,0 +1,76 @@
+﻿<UserControl 
+    x:Class="Wider.Core.Views.StatusBar"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+    xmlns:Converter="clr-namespace:Wider.Core.Converters"
+    xmlns:local="clr-namespace:Wider.Core.Views"
+    mc:Ignorable="d" 
+    d:DesignHeight="22" d:DesignWidth="480">
+    <UserControl.Resources>
+        <ResourceDictionary>
+            <Converter:NullObjectToVisibiltyConverter x:Key="nullToVisibilityConverter" />
+            <Converter:BoolToInsertModeConverter x:Key="booleanToInsertModeConverter" />
+            <BooleanToVisibilityConverter x:Key="booleanToVisibilityConverter" />
+        </ResourceDictionary>
+    </UserControl.Resources>
+    <StatusBar x:Name="WiderStatusBar" Background="{Binding Path=Background}">
+        <StatusBar.ItemsPanel>
+            <ItemsPanelTemplate>
+                <Grid>
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="3*" />
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="30" />
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="*" />
+                    </Grid.ColumnDefinitions>
+                </Grid>
+            </ItemsPanelTemplate>
+        </StatusBar.ItemsPanel>
+        <!--Status message-->
+        <StatusBarItem>
+            <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=Text, UpdateSourceTrigger=PropertyChanged}" />
+        </StatusBarItem>
+        <!--Progress bar-->
+        <StatusBarItem Grid.Column="1" HorizontalContentAlignment="Stretch" Visibility="{Binding Path=ShowProgressBar, Converter={StaticResource ResourceKey=booleanToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
+            <ProgressBar Width="Auto" 
+                         Maximum="{Binding Path=ProgressMaximum, UpdateSourceTrigger=PropertyChanged}" 
+                         Value="{Binding Path=ProgressValue, UpdateSourceTrigger=PropertyChanged}" />
+        </StatusBarItem>
+        <!--Icon animation-->
+        <StatusBarItem Grid.Column="2" 
+                       Visibility="{Binding Path=AnimationImage, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
+            <ContentControl Content="{Binding Path=AnimationImage, UpdateSourceTrigger=PropertyChanged}" />
+        </StatusBarItem>
+        <!--Line number-->
+        <StatusBarItem Grid.Column="3" 
+                       Visibility="{Binding Path=LineNumber, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="Ln " />
+                <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=LineNumber, UpdateSourceTrigger=PropertyChanged}" />
+            </StackPanel>
+        </StatusBarItem>
+        <!--Column number-->
+        <StatusBarItem Grid.Column="4" Visibility="{Binding Path=ColPosition, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="Col " />
+                <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=ColPosition}" />
+            </StackPanel>
+        </StatusBarItem>
+        <!--Character number-->
+        <StatusBarItem Grid.Column="5" Visibility="{Binding Path=CharPosition, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="Ch " />
+                <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=StatusBar.CharPosition, UpdateSourceTrigger=PropertyChanged}" />
+            </StackPanel>
+        </StatusBarItem>
+        <!--Insert-->
+        <StatusBarItem Grid.Column="6" HorizontalAlignment="Right" Visibility="{Binding Path=InsertMode, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
+            <TextBlock Foreground="{Binding Path=Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=StatusBar.InsertMode, Converter={StaticResource ResourceKey=booleanToInsertModeConverter}, UpdateSourceTrigger=PropertyChanged}" />
+        </StatusBarItem>
+    </StatusBar>
+</UserControl>
