+++ Framework/Wider.Core.Wpf/Views/StatusBar.xaml.cs
@@ -1,0 +1,31 @@
+﻿using Prism.Ioc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+using Wider.Core.Services;
+
+namespace Wider.Core.Views
+{
+    /// <summary>
+    /// Interaction logic for StatusBar.xaml
+    /// </summary>
+    public partial class StatusBar : UserControl
+    {
+        public StatusBar(IContainerExtension containerExtension)
+        {
+            InitializeComponent();
+            DataContext = containerExtension.Resolve<IStatusbarService>();
+        }
+    }
+}
