+++ Modules/Wider.Shell.Metro/Views/ShellView.xaml
@@ -7,6 +7,7 @@
     xmlns:Controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
     xmlns:behaviours="clr-namespace:MahApps.Metro.Behaviours;assembly=MahApps.Metro"
     xmlns:Converter="clr-namespace:Wider.Core.Converters;assembly=Wider.Core"
+    xmlns:prism ="http://prismlibrary.com/"
     xmlns:Utils="clr-namespace:Wider.Utils;assembly=Wider.Core"
     TitleForeground="#FF717171" 
     TitleCaps="False" 
@@ -54,62 +55,10 @@
                   ItemsSource="{Binding Path=Menus, UpdateSourceTrigger=PropertyChanged}" />
             <!--Toolbar (can we improve this?)-->
             <ContentControl Content="{Binding ToolBarTray}" DockPanel.Dock="Top" />
+
             <!--Status bar-->
-            <StatusBar x:Name="WiderStatusBar" DockPanel.Dock="Bottom" Background="{Binding Path=StatusBar.Background}">
-                <StatusBar.ItemsPanel>
-                    <ItemsPanelTemplate>
-                        <Grid>
-                            <Grid.ColumnDefinitions>
-                                <ColumnDefinition Width="3*" />
-                                <ColumnDefinition Width="*" />
-                                <ColumnDefinition Width="30" />
-                                <ColumnDefinition Width="*" />
-                                <ColumnDefinition Width="*" />
-                                <ColumnDefinition Width="*" />
-                                <ColumnDefinition Width="*" />
-                            </Grid.ColumnDefinitions>
-                        </Grid>
-                    </ItemsPanelTemplate>
-                </StatusBar.ItemsPanel>
-                <!--Status message-->
-                <StatusBarItem>
-                    <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=StatusBar.Text, UpdateSourceTrigger=PropertyChanged}" />
-                </StatusBarItem>
-                <!--Progress bar-->
-                <StatusBarItem Grid.Column="1" HorizontalContentAlignment="Stretch" Visibility="{Binding Path=StatusBar.ShowProgressBar, Converter={StaticResource ResourceKey=booleanToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
-                    <ProgressBar Width="Auto" Maximum="{Binding Path=StatusBar.ProgressMaximum, UpdateSourceTrigger=PropertyChanged}" Value="{Binding Path=StatusBar.ProgressValue, UpdateSourceTrigger=PropertyChanged}" />
-                </StatusBarItem>
-                <!--Icon animation-->
-                <StatusBarItem Grid.Column="2" Visibility="{Binding Path=StatusBar.AnimationImage, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
-                    <ContentControl Content="{Binding Path=StatusBar.AnimationImage, UpdateSourceTrigger=PropertyChanged}" />
-                </StatusBarItem>
-                <!--Line number-->
-                <StatusBarItem Grid.Column="3" Visibility="{Binding Path=StatusBar.LineNumber, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
-                    <StackPanel Orientation="Horizontal">
-                        <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="Ln " />
-                        <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=StatusBar.LineNumber, UpdateSourceTrigger=PropertyChanged}" />
-                    </StackPanel>
-                </StatusBarItem>
-                <!--Column number-->
-                <StatusBarItem Grid.Column="4" Visibility="{Binding Path=StatusBar.ColPosition, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
-                    <StackPanel Orientation="Horizontal">
-                        <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="Col " />
-                        <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=StatusBar.ColPosition}" />
-                    </StackPanel>
-                </StatusBarItem>
-                <!--Character number-->
-                <StatusBarItem Grid.Column="5" Visibility="{Binding Path=StatusBar.CharPosition, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
-                    <StackPanel Orientation="Horizontal">
-                        <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="Ch " />
-                        <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=StatusBar.CharPosition, UpdateSourceTrigger=PropertyChanged}" />
-                    </StackPanel>
-                </StatusBarItem>
-                <!--Insert-->
-                <StatusBarItem Grid.Column="6" HorizontalAlignment="Right" Visibility="{Binding Path=StatusBar.InsertMode, Converter={StaticResource ResourceKey=nullToVisibilityConverter}, UpdateSourceTrigger=PropertyChanged}">
-                    <TextBlock Foreground="{Binding Path=StatusBar.Foreground, UpdateSourceTrigger=PropertyChanged}" Text="{Binding Path=StatusBar.InsertMode, Converter={StaticResource ResourceKey=booleanToInsertModeConverter}, UpdateSourceTrigger=PropertyChanged}" />
-                </StatusBarItem>
-            </StatusBar>
-            
+            <ItemsControl Name="StatusBar" DockPanel.Dock="Bottom" prism:RegionManager.RegionName="StatusBar" />
+
             <!--The magical avalon dock-->
             <avalonDock:DockingManager 
                 x:Name="dockManager" AnchorablesSource="{Binding Tools}"
